requirements.txt
datasets
pandas
sentence-transformers
faiss-cpu
openai
python-dotenv
graphviz
streamlit
PyMuPDF
matplotlib